Updated build workflows to use git version instead of conan
The workflows in the .github folder have been updated to use git version instead of conan to build the software. These changes were made in the "macos.yml", "linux.yml", "windows.yml", "release.yml", and "conan-package.yml" files. The reason for this change is to avoid potential issues with Conan and streamline the build process. This includes eliminating the use of conan-package.yml and updating the inputs for the plugin version.