Add Ninja setup to Windows workflow and remove macOS ARM64 binaries artifacts from Release workflow
In the Windows Github Actions workflow, the Ninja tool setup has been added for efficient building of source code. This enhances the build process for Windows.

In the Release workflow, the downloading of macOS ARM64 binary artifacts has been removed. This is due to the decision of no longer supporting these type of builds in our release process.