Expand README with plugin details and license notice
Added more detailed information to the README file for the CuraEngine Only Fans plugin. Besides a brief description of the plugin functionality, a notice regarding the source files' license has been included too. The notice clarifies the current absence of a license and states an intention to update it once changes are reflected from the Ultimaker repositories.