Update Conan install command for C++20 support
Removed the use of seanmiddleditch/gha-setup-ninja from windows YAML workflow as it is no longer necessary. Added a specific compiler flag to the Conan install command in the windows, Linux, and macOS YAML workflows to ensure the use of C++20 standard during the build process. This helps in keeping the codebase up to date with modern C++ standards.