Update Conan install commands and version setter
The Conan install commands in the GitHub workflows for Windows, Linux, and MacOS have been updated to specify a plugin Git version for `curaengine_onlyfans` package. The removal of `seanmiddleditch/gha-setup-ninja@master` from the Windows workflow is also part of this commit. In the conanfile.py, the version setter now checks if the version is 'main', besides being None. These changes provide more flexibility in the build process and fix an issue with the default 'main' version.