Update set_version condition in conanfile.py
The set_version condition in conanfile.py has been simplified. Previously, it was checking both "if version is not set" and "if version is 'main'". The latter check has now been removed because it was unnecessary redundancy, as "main" should not be used as a version number. Now the function only ensures a version has been set to prevent problems with