Update conan config sources and remove ninja installation
This commit updates the source of Conan configuration in the windows, linux, and macos workflow files from 'Ultimaker' to 'jellespijker'. This aligns with the team's decision to utilize a more updated and appropriate configuration for our development environment. Moreover, the ninja installation step in the windows workflow file was removed as we found out it's no longer necessary for our current build process.